Frontend wokring with barters, transactions and users

+++ Knowlify.Client/src/services/barter.ts
@@ -11,31 +11,45 @@ export const getBarter = async (id: number) => {
   return response
 }
 
-export const addBarter = async (barter: IBarter) => {
+export const addBarter = async (barter: Partial<IBarter>, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.barter.add, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(barter),
   })
   return response
 }
 
-export const updateBarter = async (barter: IBarter) => {
+export const updateBarter = async (barter: IBarter, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.barter.update, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(barter),
   })
   return response
 }
 
-export const deleteBarter = async (id: number) => {
+export const deleteBarter = async (id: number, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(`${API_ROUTES.barter.delete}/${id}`, {
     method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   })
   return response
 }
