+++ Knowlify.Client/src/services/transaction.ts
@@ -1,5 +1,5 @@
 import { API_ROUTES } from '@/constants/api'
-import { ITransaction } from '@/interfaces/transaction'
+import { ITransactionRequest, ITransaction } from '@/interfaces/transaction'
 
 export const getTransactions = async () => {
   const response = await fetch(API_ROUTES.transaction.all)
@@ -12,23 +12,36 @@ export const getTransaction = async (id: number) => {
   return response
 }
 
-export const addTransaction = async (transaction: ITransaction) => {
+export const addTransaction = async (
+  transaction: ITransactionRequest,
+  token: string,
+) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.transaction.add, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(transaction),
   })
-  const data = (await response.json()) as ITransaction
-  return data
+  return response
 }
 
-export const updateTransaction = async (transaction: ITransaction) => {
+export const updateTransaction = async (
+  transaction: ITransaction,
+  token: string,
+) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.transaction.update, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(transaction),
   })
@@ -36,10 +49,21 @@ export const updateTransaction = async (transaction: ITransaction) => {
   return data
 }
 
-export const deleteTransaction = async (id: number) => {
+export const deleteTransaction = async (id: number, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(`${API_ROUTES.transaction.delete}/${id}`, {
     method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   })
   const data = (await response.json()) as ITransaction
   return data
+}
+
+export const getLastTransactions = async () => {
+  const response = await fetch(API_ROUTES.transaction.last)
+  return response
 }
